Updated load_data.sh to check for tty
Updated load_data.sh to see if terminal running the command is tty. If
it is it will continue regulary with running the docker script to load
data. If it is not tty assuming that the person running load_data is
using Git Bash it will prepend winpty to the docker command.