EREGCSC-743 load supplementary content (#135)
* Created lamba to add in previous supp content

* Added makefile entry to build supp content

* Make github action to load supp content

* Added create to supplementary content serializer

Added a create method to CategorySerializer to be able to create
categories with nested supplementary_content and regulation sections.

Right now the nesting for creating sections doesn't work because of the
many-to-many fields. Django recommends using a set().

* Load all supp content and handle edge cases

* Add approved field to supplemental content

Return supplemental content only if it has been approved. (Default is
approved)

* Include parsed guidance for upload

* Log information about loading supplemental content

* Setup supp content path in serverless

* Refactor supp content use  content vs categories

Fetches applicable content rather than categories.
Uses serializer logic to build the finished list of categories.

* Require supp content to belong to a category

* Remove loaded supp content from git

* Add pulling supp content to the gh deploy action

Co-authored-by: David Gage <whytheplatypus@gmail.com>