Change match behavior (see commit msg)
If Act is specified (i.e. "of the Social Security Act"), then we look it up in the conversions table.
Link is created only if section exists within the specified act.

If no act is specified, we default to Social Security Act. If section isn't found within SSA,
no link is generated.