EREGCSC-1843 Automatically link to SSA in the reg text (#813)
* Add linking directly to statutes

* Precompile regex for admin page as well

* Use re.sub instead of str replace to elimianate duplicate linking

* Fix refs not linking in special text

* Include or matching, everything is one paragraph

* Remove unused paragraph stuff for now

* Fix regex not including act name

* Change match behavior (see commit msg)

If Act is specified (i.e. "of the Social Security Act"), then we look it up in the conversions table.
Link is created only if section exists within the specified act.

If no act is specified, we default to Social Security Act. If section isn't found within SSA,
no link is generated.

* Add external link icon

* Remove unneeded migration

* Improve regex

* Set recommended default to Social Security Act on admin page

* Updated tests

* Imports

* Imports again

* Import ordering

* Fix unescaped period after section

* linting