EREGCSC-2818 Improve robots.txt handling  (#1703)
* Add robots.txt allow list field to ResourcesConfiguration

* Add unit test

* Linting

* Simplify robots.txt handling in the text extractor, rewrite tests

We were trying to implement too much logic for robots.txt handling, including
the behavior of a 404 permitting crawling, that the RobotFileParser in Python
already handles for us.

* Fix schema placeholder

* Reorder widgets on ResourcesConfiguration admin page for readability

This is because the category dropdown and allow list widgets show help text under the widget,
while for the auto extract checkbox the text is aligned to the left. Better visual flow if
items similarly justified are grouped together.

* Debug message if _ignore_robots is true