Fix for FR parser bug: field required for section child_id
Because the SectionCreateSerializer was a ModelSerializer with a Meta class referencing all fields,
the FR parser was failing because the recently added "child_id" field is not included in the request
from the parser. Since this field is set on save anyway, we can simply ignore it by defining the serializer
as a "serializers.Serializer" and manually define the fields, like we do with many other serializers already.