Use asynchronous Textract for all PDF files
PDF extractor will upload the document to a preconfigured S3 bucket,
then call boto3's start_document_text_detection() method. If an SQS queue is
configured, push the original message with "job_id" set to the queue, then return early
with no message sent to eRegs.

Text extractor main handler modified so that a None response from extract() will
trigger an early return. Also modified so that if "job_id" is in config, we assume
a continuation of a previously started extraction (of some arbitrary kind) and skip
the file retrieval step. If "file_type" is in config, we assume the file type also
instead of using Magika to determine it ourselves.

If an SQS queue is not set, we will sleep 60 seconds and, using recursion, recheck
the job's status until it fails, completes, or the Lambda times out. On success,
the text is returned.