EREGCSC-2971 Enable using Textract async text extraction (#1726)
* Create infrastructure for async PDF extraction

- New SNS topic that Textract is allowed to publish to.
- New S3 bucket for PDF storage that Textract can read from, and Lambda can read/write/delete.
- Existing SQS queue subscribed to SNS topic.

* SNS topic -> FIFO SNS topic to be compatible with FIFO SQS

* Create discrete IAM role for Textract to publish to SNS

* Print event to log and exit

* Change permissions boundary id

* Prepend SNS topic name with AmazonTextract

See https://docs.aws.amazon.com/textract/latest/dg/api-async-roles.html

* Attempt using EventBridge for message passing between Textract -> S3 -> SQS

* Include group ID for FIFO

* Remove EventBridge etc., will handle Textract logic in Lambda

* Combine textract and image extractor into a single image extractor

* Use asynchronous Textract for all PDF files

PDF extractor will upload the document to a preconfigured S3 bucket,
then call boto3's start_document_text_detection() method. If an SQS queue is
configured, push the original message with "job_id" set to the queue, then return early
with no message sent to eRegs.

Text extractor main handler modified so that a None response from extract() will
trigger an early return. Also modified so that if "job_id" is in config, we assume
a continuation of a previously started extraction (of some arbitrary kind) and skip
the file retrieval step. If "file_type" is in config, we assume the file type also
instead of using Magika to determine it ourselves.

If an SQS queue is not set, we will sleep 60 seconds and, using recursion, recheck
the job's status until it fails, completes, or the Lambda times out. On success,
the text is returned.

* Add TEXTRACT_BUCKET and TEXT_EXTRACTOR_QUEUE_URL to Lambda environment vars

* Re-add text extractor ARN

* List* to ListBucket

* Fix unsupported action

* Remove debug statements from text extractor handler function

* Add SQS SendMessage permission to Lambda

* Add S3 as resource for ListBucket permission (enabling resource deletion)

* Implement delay time to allow async processing to occur

* Add new unit tests for PDF extractor

* Remove PDF fixtures, tests no longer require them

* Add sleep before recursion and update README