EREGCSC-2094-B Fix for "part"."section" format registering "part" as a section (#1770)
* Fix for "part"."section" format registering "part" as a section if conversion exists

Existing regex does not consider "123.456" to be a "part" and a "section", it only looks
for the section e.g. "§ 123", so the period being there splits the reference and the site
will link to section "123" if a link conversion exists for it.

Instead, we have to use negative lookahead to ensure that a reference that looks
"§ 123.456" is ignored entirely. These refs will be handled by an entirely separate regex.

* Factor out the negative lookahead pattern for clarity and reusability