Fix distinct and sorting; use Django ORM for generating initial query
We have been trying to combine the existing filters for content search
with the raw SQL required for the new hybrid search techniques. Attempted
to implement filters in raw SQL and that was too complex. Then attempted
to use IDs retrieved from ORM filtering to filter hybrid search, that was too slow.

This approach defines the initial filters by the ORM, then retrieves
the SQL query that would be used. We can insert that as a CTE in the
hybrid search SQL that the other CTEs use like:

SELECT id FROM content_search_contentindex WHERE id IN ( SELECT id FROM indices );

This avoids having to pass in a large (and ever growing) array of IDs into the "IN" clause.

Also, this approach is by far the fastest we've ever seen! Simple search results
locally come back in as little as 100ms.