EREGCSC-3010 Integrate semantic search with RRF into content search (#1897)
* Initial SQL for hybrid search

* filter results before performing search

* Fix pagination and improve parameter handling

* add comments, linting

* Results now distinct by resource_id and reg_text_id

* fix incomplete merge

* Add tweakable search parameters to admin panel

* Enable quoted search and use params for search strategy

* Optimize search SQL: compile Q objects to allow filter by ID without 2nd query

* Cleanup

* Fixups

* Fix distinct and sorting; use Django ORM for generating initial query

We have been trying to combine the existing filters for content search
with the raw SQL required for the new hybrid search techniques. Attempted
to implement filters in raw SQL and that was too complex. Then attempted
to use IDs retrieved from ORM filtering to filter hybrid search, that was too slow.

This approach defines the initial filters by the ORM, then retrieves
the SQL query that would be used. We can insert that as a CTE in the
hybrid search SQL that the other CTEs use like:

SELECT id FROM content_search_contentindex WHERE id IN ( SELECT id FROM indices );

This avoids having to pass in a large (and ever growing) array of IDs into the "IN" clause.

Also, this approach is by far the fastest we've ever seen! Simple search results
locally come back in as little as 100ms.

* Fix counts endpoint

* Fix rank order for keyword search

* Move headline generation to search viewset, add options to admin panel

* Add min filter for quoted search to admin panel

* Temporarily pin Django version to less than 4.2.8

This is necessary because the 4.2.8 breaks URL state for the search frontend
when (queries or) URLs are greater than 2048 characters.

* Fix eCFR parser failure: IndexedRegulationText has no extraction_error field

* Update Swagger docs; fix parameters; allow POST for counts endpoint

* Add Swagger doc for chunk update endpoints

* Attempt to fix timeout during content_search/0011 migration

* Add embedding generation code

* linting

* remove debug statement